Worked

+++ Classes.cs
@@ -12,5 +12,5 @@ public class Grade
 {
     public string Name { get; set; }
     public int Rank { get; set; } // Lower Rank = Higher Grade
-    public int PercentageToAssign { get; set; } 
+    public double PercentageToAssign { get; set; } // It has to be double in order to work
 }
